@@ -10,19 +10,19 @@
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/adapter-auto": "^7.0.1",
+    "@sveltejs/kit": "^2.53.0",
+    "@sveltejs/vite-plugin-svelte": "^5.1.1",
     "autoprefixer": "^10.4.21",
     "postcss": "^8.5.6",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.0",
+    "svelte": "^5.53.2",
+    "svelte-check": "^4.4.3",
     "tailwindcss": "^3.4.18",
     "typescript": "^5.0.0",
-    "vite": "^5.0.3"
+    "vite": "^6.4.1"
   },
   "type": "module",
   "dependencies": {
-    "@sveltejs/adapter-node": "^4.0.1"
+    "@sveltejs/adapter-node": "^5.5.3"
   }
 }
